+++ frontend/components/screens/group/requests/invitations/Invitations.module.scss
@@ -1,0 +1,3 @@
+.invitation_container {
+	@apply flex flex-col gap-5;
+}
